Initial Commit i guess
just getting some basics down. right now I am able to crawl through the folders and locate the subtitle folder. the next step is locating the appropriate subtitle file and then renaming and moving it. I think once I'm past this hump of running regex on a list I'll be able to get the basics up